Add metadata to event scrape output
