+++ openstates/scrape/event.py
@@ -212,3 +212,6 @@ class Event(BaseModel, SourceMixin, AssociatedLinkMixin, LinkMixin):
             date=date,
             classification=classification,
         )
+
+    def pre_save(self, jurisdiction):
+        self.add_scrape_metadata(jurisdiction)
